+++ weed/filer2/reader_at.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/chrislusf/seaweedfs/weed/glog"
 	"github.com/chrislusf/seaweedfs/weed/pb/filer_pb"
-	"github.com/chrislusf/seaweedfs/weed/pb/pb_cache"
 	"github.com/chrislusf/seaweedfs/weed/util"
+	"github.com/chrislusf/seaweedfs/weed/util/pb_cache"
 	"github.com/chrislusf/seaweedfs/weed/wdclient"
 )
